+++ src/main/java/xyz/teamcarrot/myct/cart/CartController.java
@@ -1,4 +1,4 @@
-package xyz.teamcarrot.myct.goods;
+package xyz.teamcarrot.myct.cart;
 
 import java.util.List;
