+++ src/main/resources/xyz/teamcarrot/myct/cart/CartMapper.xml
@@ -1,24 +1,24 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="xyz.teamcarrot.myct.goods.GoodsMapper">
+<mapper namespace="xyz.teamcarrot.myct.cart.CartMapper">
 
-	<select id="getlist" resultType="xyz.teamcarrot.myct.goods.CartVO">
+	<select id="getlist" resultType="xyz.teamcarrot.myct.cart.CartVO">
 		SELECT *,(g.price+o.price_updown) AS final_price FROM tb_cart c
 		INNER JOIN tb_goods g ON c.goods_no=g.goods_no
 		INNER JOIN tb_goods_option o ON c.option_no=o.option_no
 	</select>
 	
-	<insert id="insert" parameterType="xyz.teamcarrot.myct.goods.CartVO">
+	<insert id="insert" parameterType="xyz.teamcarrot.myct.cart.CartVO">
 		INSERT INTO tb_cart (goods_count, cart_add_date, goods_no, member_no, option_no)
 		VALUES (#{goods_count}, #{cart_add_date}, #{goods_no}, #{member_no}, #{option_no})
 	</insert>
 	
-	<delete id="delete" parameterType="xyz.teamcarrot.myct.goods.CartVO">
+	<delete id="delete" parameterType="xyz.teamcarrot.myct.cart.CartVO">
 		DELETE FROM tb_cart
 		WHERE cart_no = #{cart_no}
 	</delete>
 	
-	<update id="update" parameterType="xyz.teamcarrot.myct.goods.CartVO">
+	<update id="update" parameterType="xyz.teamcarrot.myct.cart.CartVO">
         UPDATE tb_cart
         SET goods_count = #{goods_count}
         WHERE cart_no = #{cart_no}
